+++ zobrist.h
@@ -1,20 +1,22 @@
 #pragma once
 
+#include <stdint.h>
+
 #include "game.h"
 #include "list.h"
 
 #define HASH_TABLE_SIZE ((int) 1e6 + 3)  // choose a large prime number
 
-extern unsigned long long zobrist_table[N_GRIDS][2];
+extern uint64_t zobrist_table[N_GRIDS][2];
 
 typedef struct {
-    unsigned long long key;
+    uint64_t key;
     int score;
     int move;
     struct hlist_node ht_list;
 } zobrist_entry_t;
 
 void zobrist_init(void);
-zobrist_entry_t *zobrist_get(unsigned long long key);
-void zobrist_put(unsigned long long key, int score, int move);
+zobrist_entry_t *zobrist_get(uint64_t key);
+void zobrist_put(uint64_t key, int score, int move);
 void zobrist_clear(void);
